FEMEN Android app
---

This is an Android project I started after contacting Inna if they would
like to have FEMEN on Androids.

She said yes, but also that she had something other in mind.

Unfortunately I never heard of any other plans, so I'm releasing this
into the wild as I don't have very much time to work on it any further.

---

Basically it's a compilation of different ressources you find on the
web:

* FEMEN.org
* a Facebook account (which is shut down often and changes constantly)
* their Twitter
* their online shop
* their PayPal account (which never worked...)

---

It's working as an app though, maybe someone wants to develop this any
further or compilate it for themselves.

May the fourth be with you!

---

Sincerely

Signed-off-by: c1ph4 <sascha.schroeder@aethyx.com>